refactor: ProductAdmin 섹션 UI 컴포넌트 분리
