+++ src/basic/main.tsx
@@ -1,10 +1,5 @@
-import * as React from "react";
 import * as ReactDOM from "react-dom/client";
 
 import App from "@/basic/App";
 
-ReactDOM.createRoot(document.getElementById("root")!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
+ReactDOM.createRoot(document.getElementById("root")!).render(<App />);
